+++ spec/rubocop/cop/style/line_end_concatenation_spec.rb
@@ -46,6 +46,13 @@ describe Rubocop::Cop::Style::LineEndConcatenation do
     expect(cop.offenses).to be_empty
   end
 
+  it 'accepts string concat at line end when % literals are involved' do
+    inspect_source(cop,
+                   ['top = %(test) +',
+                    '"top"'])
+    expect(cop.offenses).to be_empty
+  end
+
   it 'autocorrects by replacing + with \\' do
     corrected = autocorrect_source(cop,
                                    ['top = "test" +',